+++ js/screens/change-screen.js
@@ -8,11 +8,11 @@ const MAX_QUESTIONS = 10;
 
 export const changeScreen = (state) => {
   if (state.notes < 0) {
-    resultScreen(state);
+    renderScreen(resultScreen(state));
   } else if (state.time < 0) {
-    resultScreen(state);
+    renderScreen(resultScreen(state));
   } else if (state.level > MAX_QUESTIONS) {
-    resultScreen(state);
+    renderScreen(resultScreen(state));
   } else if (levels[state.level].type === `artist`) {
     renderScreen(artistScreen(state, changeScreen));
   } else if (levels[state.level].type === `genre`) {
